% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-load_modular.R
\name{load_modular}
\alias{load_modular}
\title{Load Modular R Package}
\usage{
load_modular(path = ".", ...)
}
\arguments{
\item{path}{The path to the root directory of the modular R package.}

\item{...}{Additional parameters passed to devtools::load_all().}
}
\value{
The function is designed to be used for its side effects. It invisibly returns NULL.
}
\description{
This function loads and documents a modular R package. It assumes that the package has
been set up for modular use with the \code{use_modular_package} function.
}
\details{
This function performs the following steps:
\enumerate{
\item Checks if the required directories ('workingFiles' and 'Rbackup') exist. If not,
it stops execution with an error message suggesting the use of 'use_modular_package'.
\item Removes existing backup files in the 'Rbackup' directory.
\item Copies the current R files to the 'Rbackup' directory.
\item Removes the current R files if the copy to 'Rbackup' was successful.
\item Retrieves R working files from the 'workingFiles' directory.
\item Modifies the file names by appending the paths and copies them into the 'R' directory.
\item Loads all R files and documents the package using devtools.
}
}
\examples{
\dontrun{
# Load the modular package with the default path
load_modular()

# Load the modular package with a custom path
load_modular('/path/to/package')
}

}
