% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-use_modular_package.R
\name{use_modular_package}
\alias{use_modular_package}
\title{Set Up Modular Package}
\usage{
use_modular_package(path = here::here())
}
\arguments{
\item{path}{The path where the package directories should be created.}
}
\value{
The function is designed to be used for its side effects. It invisibly returns NULL.
}
\description{
This function sets up the necessary directories for a modular R package, in the
style of a function from usethis.
}
\details{
This function creates two directories, 'workingFiles' and 'Rbackup', in
the specified path. If these directories do not exist, it creates them. Additionally,
it adds 'workingFiles' and 'Rbackup' to the build ignore list using \code{usethis::use_build_ignore}.
}
\examples{
\dontrun{
# Use the function with the default path
use_modular_package()

# Use the function with a custom path
use_modular_package('/path/to/package')
}

}
